requirements.txt
numpy==1.22
requests==2.27.1
eth-brownie==1.18
